backend/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
python-multipart>=0.0.9

# AI / ML
transformers>=4.40.0
# CPU-only torch (smaller image, sufficient for CLIP inference)
torch>=2.3.0
Pillow>=10.3.0

# Background removal
rembg>=2.0.57
onnxruntime>=1.17.0

# 3D generation
gradio_client>=0.16.0

python-dotenv>=1.0.1
